Fixed F# samples

+++ samples/fsharp/getting-started/BinaryClassification_SentimentAnalysis/Program.fs
@@ -14,16 +14,16 @@ type SentimentData() =
     member val SentimentText: string = "" with get, set
 
     [<Column("1", name="Label")>]
-    member val  Sentiment : double = 0.0 with get, set
+    member val Sentiment: float32 = 0.0f with get, set
 
 type SentimentPrediction() =
     [<ColumnName("PredictedLabel")>]
-    member val  Sentiment : bool = false with get, set
+    member val Sentiment: bool = false with get, set
 
 let sentiments = 
-   [| SentimentData(SentimentText = "Contoso's 11 is a wonderful experience", Sentiment = 1.0)
-      SentimentData(SentimentText = "The acting in this movie is very bad", Sentiment = 0.0)
-      SentimentData(SentimentText = "Joe versus the Volcano Coffee Company is a great film.", Sentiment = 1.0) |]
+   [| SentimentData(SentimentText = "Contoso's 11 is a wonderful experience", Sentiment = 1.0f)
+      SentimentData(SentimentText = "The acting in this movie is very bad", Sentiment = 0.0f)
+      SentimentData(SentimentText = "Joe versus the Volcano Coffee Company is a great film.", Sentiment = 1.0f) |]
 
 let AppPath = Path.Combine(__SOURCE_DIRECTORY__, "../../../..")
 let TrainDataPath = Path.Combine(AppPath, "datasets", "sentiment-imdb-train.txt")
